// apps/web/src/gen/api/admin/model/user.ts
/**
 * Admin
 * Admin service is responsible for User management and authentication. Also it provides static data i.e list of Skills and Countries 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Skill } from './skill';
import { Country } from './country';


export interface User { 
    username?: string;
    passwordHash?: string;
    id: string;
    googleId: string;
    dateCreated: string;
    firstName: string;
    lastName: string;
    birthDate: string;
    gender: User.GenderEnum;
    experience: number;
    roles: Array<User.RolesEnum>;
    country: Country;
    skills: Array<Skill>;
}
export namespace User {
    export type GenderEnum = 'male' | 'female';
    export const GenderEnum = {
        Male: 'male' as GenderEnum,
        Female: 'female' as GenderEnum
    };
    export type RolesEnum = 'admin' | 'developer' | 'crowdworker';
    export const RolesEnum = {
        Admin: 'admin' as RolesEnum,
        Developer: 'developer' as RolesEnum,
        Crowdworker: 'crowdworker' as RolesEnum
    };
}